Fix: parity review site shows no pairs due to process.exit on import
Extract mapJsToPython into a side-effect-free path-mapping.ts module.
build-parity-review-site.ts now imports from there directly, avoiding
the top-level process.exit(1) in check-python-sync.ts that killed the
build before results.json was written.
